requirements.txt
Adafruit-Blinka==8.45.2
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.1.1
adafruit-circuitpython-neopixel==6.3.11
adafruit-circuitpython-pixelbuf==2.0.4
adafruit-circuitpython-requests==4.1.3
adafruit-circuitpython-typing==1.10.3
Adafruit-PlatformDetect==3.71.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
numpy==2.0.0
pyftdi==0.55.4
pyserial==3.5
pyusb==1.2.1
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
sysv-ipc==1.1.0
typing_extensions==4.12.2
